Send chat event messages
If the body of the posted move contains a 'chatEvent', it is sent to the
chat module in the room defined by the game's type and players. (a room
has a type and an array of users, a game has a type and an array of
players, so we just match those two)

When the game is over, the 'gameover' event is also added to the chat,
with the following format.

        "gameover:player1,player2,playerX:score1,score2,scoreX"